before:
`del ................................ 0.000 ms      0.009 ms    ±0.018  x10000`
after:
`del ................................ 0.000 ms      0.001 ms    ±0.001  x10000`

by not copying entry (assigning to tmp var)